spectrogram: bugfix

+++ spectrogram.py
@@ -145,7 +145,7 @@ class Spectrogram(Snuffling):
         p = None
 
         ncols = int(len(nslcs) / 5 + 1)
-        nrows = (len(nslcs)-1) / ncols + 1
+        nrows = int((len(nslcs)-1) / ncols + 1)
 
         tmin = min(times)
         tmax = max(times)
